# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: API Activos Informáticos
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    x-swagger-router-controller: "User"
    get:
      tags:
      - "User"
      summary: "Consulta de usuarios"
      description: "Listado de todos los usuarios autorizados a utlizar el servicio"
      operationId: "usersGet"
      parameters: []
      responses:
        200:
          description: "Listado de usuarios"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    post:
      tags:
      - "User"
      summary: "Crear usuario"
      description: "Dar de alta un usario dandole asi permiso para utlizar el servicio"
      operationId: "usersPost"
      parameters:
      - in: "body"
        name: "user"
        description: "Usuario a dar de alta en el sistema"
        required: true
        schema:
          $ref: "#/definitions/NewUser"
      responses:
        201:
          description: "Usuario creado"
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NewResource"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
  /users/{id}:
    x-swagger-router-controller: "User"
    get:
      tags:
      - "User"
      summary: "Usuario"
      description: "Usuario por identificador\n"
      operationId: "usersIdGet"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del usuario."
        required: true
        type: "string"
      responses:
        200:
          description: "Usuario por id."
          schema:
            $ref: "#/definitions/User"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    put:
      tags:
      - "User"
      summary: "Usuario"
      description: "Usuario por identificador\n"
      operationId: "usersIdPut"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del usuario."
        required: true
        type: "string"
      - in: "body"
        name: "user"
        description: "Datos de usuario para actualizar"
        required: true
        schema:
          $ref: "#/definitions/UpdateUser"
      responses:
        200:
          description: "Usuario actualizado."
          schema:
            $ref: "#/definitions/Notification"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    delete:
      tags:
      - "User"
      summary: "Usuario"
      description: "Usuario por identificador\n"
      operationId: "usersIdDelete"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del usuario."
        required: true
        type: "string"
      responses:
        200:
          description: "Usuario borrado."
          schema:
            $ref: "#/definitions/Notification"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
  /assetTypes:
    x-swagger-router-controller: "AssetType"
    get:
      tags:
      - "AssetTypes"
      summary: "Consulta de tipos de activos informáticos"
      description: "Listado de todos los tipos de activos informáticos disponibles"
      operationId: "atGet"
      parameters: []
      responses:
        200:
          description: "Listado de tipos de activos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AssetType"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    post:
      tags:
      - "AssetTypes"
      summary: "Crear tipo de activo"
      description: "Definir un nuevo tipo de activo a partir del cual se instanciarán nuevos activos informáticos."
      operationId: "atPost"
      parameters:
      - in: "body"
        name: "assetType"
        description: "Definicion del nuevo tipo de activo"
        required: true
        schema:
          $ref: "#/definitions/NewAssetType"
      responses:
        201:
          description: "Tipo de activo creado"
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NewResource"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
  /assetTypes/{id}:
    x-swagger-router-controller: "AssetType"
    get:
      tags:
      - "AssetTypes"
      summary: "Tipo de activo"
      description: "Tipo de activo por identificador\n"
      operationId: "atIdGet"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del tipo de activo."
        required: true
        type: "string"
      responses:
        200:
          description: "Tipo de activo por id."
          schema:
            $ref: "#/definitions/AssetType"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    put:
      tags:
        - "AssetTypes"
      summary: "Actualizacion de un tipo de activo informático"
      description: "Actualiza los datos asociados a un tipo de activo informático\n"
      operationId: "atIdPut"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del tipo de activo."
        required: true
        type: "string"
      - in: "body"
        name: "assetType"
        description: "Nuevos datos del activo"
        required: true
        schema:
          $ref: "#/definitions/UpdateAssetType"
      responses:
        200:
          description: "Activo actualizado."
          schema:
            $ref: "#/definitions/Notification"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
  /assets:
    x-swagger-router-controller: "Asset"
    get:
      tags:
        - "Asset"
      summary: "Consulta de activos"
      description: "Listado de todos los activos registrados en el sistema"
      operationId: "assetsGet"
      parameters: []
      responses:
        200:
          description: "Listado de activos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Asset"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    post:
      tags:
        - "Asset"
      summary: "Registrar un nuevo activo"
      description: "Dar de alta un nuevo activo en el sistema"
      operationId: "assetsPost"
      parameters:
        - in: "body"
          name: "asset"
          description: "Activo informático a regristrar en el sistema"
          required: true
          schema:
            $ref: "#/definitions/NewAsset"
      responses:
        201:
          description: "Activo creado"
          schema:
            $ref: "#/definitions/NewResource"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"

  /assetTypes/{typeId}/assets:
    x-swagger-router-controller: "Asset"
    get:
      tags:
        - "Asset"
      summary: "Consulta de activos informáticos por tipo"
      description: "Listado de todos los activos informáticos pertenecientes al tipo provisto"
      operationId: "assetsGetByType"
      parameters:
      - name: "typeId"
        in: "path"
        description: "tipo del activo."
        required: true
        type: "string"
      responses:
        200:
          description: "Listado de activos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AssetByType"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"

  /assets/{id}:
    x-swagger-router-controller: "Asset"
    get:
      tags:
        - "Asset"
      description: "Activo por identificador\n"
      operationId: "assetsIdGet"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del activo."
        required: true
        type: "string"
      responses:
        200:
          description: "Activo por id."
          schema:
            $ref: "#/definitions/Asset"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    put:
      tags:
        - "Asset"
      summary: "Actualizacion de un activo informático"
      description: "Actualiza los datos asociados a un activo informático\n"
      operationId: "assetIdPut"
      parameters:
      - name: "id"
        in: "path"
        description: "identificador del activo."
        required: true
        type: "string"
      - in: "body"
        name: "asset"
        description: "Nuevos datos del activo"
        required: true
        schema:
          $ref: "#/definitions/UpdateAsset"
      responses:
        200:
          description: "Activo actualizado."
          schema:
            $ref: "#/definitions/Notification"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"
    delete:
      tags:
        - "Asset"
      summary: "Eliminar activo"
      description: "Elimina lógicamente un activo"
      operationId: "assetIdDelete"
      parameters:
        - name: "id"
          in: "path"
          description: "identificador del activo."
          required: true
          type: "string"
      responses:
        200:
          description: "Activo borrado."
          schema:
            $ref: "#/definitions/Notification"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Notification"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Notification"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  User:
    type: "object"
    required:
      - _id
      - name
    properties:
      _id:
        type: "string"
        description: "identificador único del usuario."
      name:
        type: "string"
        description: "Nombre de usuario."
      comment:
        type: "string"
        description: "Comentario acerca del usuario."
      deleted:
        type: "boolean"
        description: "Indica si se trata de un usuario ya eliminado o no."
  UpdateUser:
    required:
      - comment
    type: "object"
    properties:
      comment:
        type: "string"
        description: "Comentario acerca del usuario."
  NewUser:
    type: "object"
    required:
      - name
    properties:
      name:
        type: "string"
        description: "Nombre de usuario."
      comment:
        type: "string"
        description: "Comentario acerca del usuario."
  AssetType:
    type: "object"
    required:
      - _id
      - name
      - properties
    properties:
      _id:
        type: "string"
        description: "identificador único del tipo de activo."
      name:
        type: "string"
        description: "Nombre del tipo de activo."
      comment:
        type: "string"
        description: "Comentario acerca del tipo de activo."
      properties:
        type: "array"
        description: "Lista de propiedades del tipo de activo."
        items:
          $ref: "#/definitions/Property"

  UpdateAssetType:
    type: "object"
    required:
      - name
      - properties
    properties:
      name:
        type: "string"
        description: "Nombre del tipo de activo."
      comment:
        type: "string"
        description: "Comentario acerca del tipo de activo."
      properties:
        type: "array"
        description: "Lista de propiedades del tipo de activo."
        items:
          $ref: "#/definitions/Property"
  Property:
    type: "object"
    required:
      - name
      - type
      - required
    properties:
      name:
        type: "string"
        description: nombre de la propiedad del activo
      type:
        type: "string"
        description: tipo de dato de la propiedad
      required:
        type: "boolean"
        description: indica si la propiedad es requerida o no
      comment:
        type: "string"
        description: breve descripcion acerca de la propiedad
  Asset:
    type: "object"
    properties:
      typeId:
        type: "string"
        required:
          - typeId
        description: id del tipo de activo
      id:
        type: "string"
        description: "identificador único"
      deleted:
        type: "boolean"
        description: "indica si el activo se encuentra borrado"
      anyProperty:
        type: ["string", "integer", "number", "boolean"]
        description: "Propiedad y valor. Pueden agregarse cuantas propiedades sean necesarias y del tipo que sean necesarias."
  UpdateAsset:
    type: "object"
    properties:
      anyProperty:
        type: ["string", "integer", "number", "boolean"]
        description: "Propiedad y valor. Pueden agregarse cuantas propiedades sean necesarias y del tipo que sean necesarias."
  AssetByType:
    type: "object"
    properties:
      id:
        type: "string"
        description: "identificador único"
      anyProperty:
        type: ["string", "integer", "number", "boolean"]
        description: "Propiedad y valor. Pueden agregarse cuantas propiedades sean necesarias y del tipo que sean necesarias."
  NewAsset:
    type: "object"
    properties:
      typeId:
        type: "string"
        description: id del tipo de activo
      anyProperty:
        type: ["string", "integer", "number", "boolean"]
        description: "Propiedad y valor. Pueden agregarse cuantas propiedades sean necesarias y del tipo que sean necesarias."
  NewAssetType:
    type: "object"
    required:
      - name
      - properties
    properties:
      name:
        type: "string"
        description: "Nombre del tipo de activo."
      comment:
        type: "string"
        description: "Comentario acerca del tipo de activo."
      properties:
        type: "array"
        description: "Lista de propiedades del tipo de activo."
        items:
          $ref: "#/definitions/Property"
  NewResource:
    type: "object"
    required:
      - id
    properties:
      id:
        type: string
  Notification:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"